--- src/main/java/com/kr/kopoctc/planningBridge/admin/domain/Department.java
@@ -1,15 +1,0 @@
-package com.kr.kopoctc.planningBridge.admin.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-public class Department {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Long departmentPK;
-
-    @Column
-    private String name;
-}
