+++ src/main/java/com/kr/kopoctc/planningBridge/admin/entity/Department.java
@@ -1,0 +1,22 @@
+package com.kr.kopoctc.planningBridge.admin.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@Entity
+public class Department {
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    private Long departmentPK;
+
+    @OneToMany(mappedBy = "department")
+    private List<User> users;
+
+    @Column
+    private String name;
+}
