+++ src/main/java/com/kr/kopoctc/planningBridge/admin/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.kr.kopoctc.planningBridge.admin.repository;
 
-import com.kr.kopoctc.planningBridge.admin.domain.User;
+import com.kr.kopoctc.planningBridge.admin.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
