+++ src/main/java/com/kr/kopoctc/planningBridge/admin/service/UserService.java
@@ -1,7 +1,7 @@
 package com.kr.kopoctc.planningBridge.admin.service;
 
 
-import com.kr.kopoctc.planningBridge.admin.domain.User;
+import com.kr.kopoctc.planningBridge.admin.entity.User;
 import com.kr.kopoctc.planningBridge.admin.dto.UserDTO;
 
 public interface UserService {
