+++ src/main/java/com/kr/kopoctc/planningBridge/common/UserType.java
@@ -1,4 +1,4 @@
-package com.kr.kopoctc.planningBridge.admin.enums;
+package com.kr.kopoctc.planningBridge.common;
 
 public enum UserType {
     POSITION,
