--- src/main/java/com/kr/kopoctc/planningBridge/project/repository/GuestRequestRepository.java
@@ -1,8 +1,0 @@
-package com.kr.kopoctc.planningBridge.project.repository;
-
-import com.kr.kopoctc.planningBridge.project.entity.GuestRequestEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GuestRequestRepository extends JpaRepository<GuestRequestEntity, String> {
-
-}
