--- src/main/java/com/kr/kopoctc/planningBridge/project/repository/ProjectHistoryRepository.java
@@ -1,8 +1,0 @@
-package com.kr.kopoctc.planningBridge.project.repository;
-
-import com.kr.kopoctc.planningBridge.project.entity.ProjectHistoryEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProjectHistoryRepository extends JpaRepository<ProjectHistoryEntity, String> {
-
-}
