--- src/main/java/com/kr/kopoctc/planningBridge/project/repository/ProjectTeamRepository.java
@@ -1,7 +1,0 @@
-package com.kr.kopoctc.planningBridge.project.repository;
-
-import com.kr.kopoctc.planningBridge.project.entity.ProjectTeamEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProjectTeamRepository extends JpaRepository<ProjectTeamEntity, String> {
-}
