--- src/main/java/com/kr/kopoctc/planningBridge/project/service/GuestRequestService.java
@@ -1,4 +1,0 @@
-package com.kr.kopoctc.planningBridge.project.service;
-
-public interface GuestRequestService {
-}
