--- src/main/java/com/kr/kopoctc/planningBridge/project/service/ProjectHistoryService.java
@@ -1,4 +1,0 @@
-package com.kr.kopoctc.planningBridge.project.service;
-
-public interface ProjectHistoryService {
-}
