--- src/main/java/com/kr/kopoctc/planningBridge/project/service/ProjectTeamService.java
@@ -1,4 +1,0 @@
-package com.kr.kopoctc.planningBridge.project.service;
-
-public interface ProjectTeamService {
-}
