--- src/main/java/com/kr/kopoctc/planningBridge/project/testController.java
@@ -1,4 +1,0 @@
-package com.kr.kopoctc.planningBridge.project;
-
-public class testController {
-}
