--- src/main/java/com/kr/kopoctc/planningBridge/task/test2.java
@@ -1,4 +1,0 @@
-package com.kr.kopoctc.planningBridge.task;
-
-public class test2 {
-}
