+++ src/main/resources/templates/admin/registerUser.mustache
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>User Registration Form</title>
-    <link rel="stylesheet" href="/css/registerUser.css">
+    <link rel="stylesheet" href="/css/admin/registerUser.css">
 </head>
 <body>
 <div class="container">
