+++ src/test/resources/application-test.yml
@@ -7,7 +7,7 @@ spring:
     lazy-initialization: true
   jpa:
     hibernate:
-      ddl-auto: create-drop
+      ddl-auto: update
     show-sql: true
     properties:
       hibernate:
